feat(commons): add esmodule support
