+++ packages/commons/jest.config.cjs
@@ -4,6 +4,9 @@ module.exports = {
     color: 'red',
   },
   preset: 'ts-jest',
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+  },
   transform: {
     '^.+\\.ts?$': 'ts-jest',
   },
@@ -13,7 +16,7 @@ module.exports = {
   roots: ['<rootDir>/src', '<rootDir>/tests'],
   testPathIgnorePatterns: ['/node_modules/'],
   testEnvironment: 'node',
-  coveragePathIgnorePatterns: ['/node_modules/'],
+  coveragePathIgnorePatterns: ['/node_modules/', 'src/types/index.ts'],
   coverageThreshold: {
     global: {
       statements: 100,
