--- packages/commons/src/awsSdk/index.ts
@@ -1,2 +1,0 @@
-export { addUserAgentMiddleware } from './userAgentMiddleware';
-export { isSdkClient } from './utils';
