+++ packages/commons/src/awsSdkUtils.ts
@@ -1,6 +1,5 @@
-import { PT_VERSION } from '../version';
-import { isSdkClient } from './utils';
-import type { MiddlewareArgsLike } from '../types/awsSdk';
+import { PT_VERSION } from './version.js';
+import type { MiddlewareArgsLike, SdkClient } from './types/awsSdk.js';
 
 /**
  * @internal
@@ -12,6 +11,28 @@ const middlewareOptions = {
   name: 'addPowertoolsToUserAgent',
   tags: ['POWERTOOLS', 'USER_AGENT'],
 };
+
+/**
+ * @internal
+ * Type guard to check if the client provided is a valid AWS SDK v3 client
+ */
+const isSdkClient = (client: unknown): client is SdkClient =>
+  typeof client === 'object' &&
+  client !== null &&
+  'send' in client &&
+  typeof client.send === 'function' &&
+  'config' in client &&
+  client.config !== undefined &&
+  typeof client.config === 'object' &&
+  client.config !== null &&
+  'middlewareStack' in client &&
+  client.middlewareStack !== undefined &&
+  typeof client.middlewareStack === 'object' &&
+  client.middlewareStack !== null &&
+  'identify' in client.middlewareStack &&
+  typeof client.middlewareStack.identify === 'function' &&
+  'addRelativeTo' in client.middlewareStack &&
+  typeof client.middlewareStack.addRelativeTo === 'function';
 
 /**
  * @internal
@@ -54,4 +75,4 @@ const addUserAgentMiddleware = (client: unknown, feature: string): void => {
   }
 };
 
-export { customUserAgentMiddleware, addUserAgentMiddleware };
+export { customUserAgentMiddleware, addUserAgentMiddleware, isSdkClient };
