--- packages/commons/src/config/index.ts
@@ -1,2 +1,0 @@
-export * from './ConfigService';
-export * from './EnvironmentVariablesService';
