+++ packages/commons/src/guards.ts
@@ -50,10 +50,3 @@ const isString = (value: unknown): value is string => {
 };
 
 export { isRecord, isString, isTruthy, isNullOrUndefined };
-
-type JSONPrimitive = string | number | boolean | null | undefined;
-type JSONValue = JSONPrimitive | JSONObject | JSONArray;
-type JSONObject = { [key: number | string]: JSONValue };
-type JSONArray = Array<JSONValue>;
-
-export type { JSONPrimitive, JSONValue, JSONObject, JSONArray };
