+++ packages/commons/src/index.ts
@@ -1,8 +1,13 @@
-export * from './utils/lambda';
-export * from './Utility';
-export * from './config';
-export * as ContextExamples from './samples/resources/contexts';
-export * as Events from './samples/resources/events';
-export * from './types/middy';
-export * from './types/utils';
-export * from './awsSdk';
+export { isRecord, isString, isTruthy, isNullOrUndefined } from './guards.js';
+export { Utility } from './Utility.js';
+export { EnvironmentVariablesService } from './config/EnvironmentVariablesService.js';
+export * as ContextExamples from './samples/resources/contexts/index.js';
+export * as Events from './samples/resources/events/index.js';
+export { addUserAgentMiddleware, isSdkClient } from './awsSdkUtils.js';
+export { cleanupMiddlewares } from './middleware/cleanupMiddlewares.js';
+export {
+  TRACER_KEY,
+  LOGGER_KEY,
+  METRICS_KEY,
+  IDEMPOTENCY_KEY,
+} from './middleware/constants.js';
