--- packages/commons/src/middleware/index.ts
@@ -1,2 +1,0 @@
-export * from './cleanupMiddlewares';
-export * from './constants';
