+++ packages/commons/src/samples/resources/contexts/index.ts
@@ -1,1 +1,1 @@
-export * from './hello-world';
+export * from './hello-world.js';
