+++ packages/commons/src/types/json.ts
@@ -1,0 +1,6 @@
+type JSONPrimitive = string | number | boolean | null | undefined;
+type JSONValue = JSONPrimitive | JSONObject | JSONArray;
+type JSONObject = { [key: number | string]: JSONValue };
+type JSONArray = Array<JSONValue>;
+
+export type { JSONPrimitive, JSONValue, JSONObject, JSONArray };
