--- packages/commons/src/utils/lambda/index.ts
@@ -1,1 +1,0 @@
-export * from './LambdaInterface';
