+++ packages/commons/tests/unit/awsSdkUtils.test.ts
@@ -1,6 +1,6 @@
-import { addUserAgentMiddleware, isSdkClient } from '../../src/awsSdk';
-import { PT_VERSION as version } from '../../src/version';
-import { customUserAgentMiddleware } from '../../src/awsSdk/userAgentMiddleware';
+import { addUserAgentMiddleware, isSdkClient } from '../../src/index.js';
+import { customUserAgentMiddleware } from '../../src/awsSdkUtils.js';
+import { PT_VERSION as version } from '../../src/version.js';
 
 describe('Helpers: awsSdk', () => {
   describe('Function: userAgentMiddleware', () => {
